+++ R/mirai-package.R
@@ -1,13 +1,13 @@
 #' mirai: Minimalist Async Evaluation Framework for R
 #'
-#' An extremely simple and lightweight method for concurrent / parallel code
-#'     execution, built on the nanonext R package and NNG (Nanomsg Next Gen).
-#'
-#'     A mirai is a future in Japanese.
+#' Extremely simple and lightweight method for concurrent / parallel code
+#'     execution, built on 'nanonext' and 'NNG' (Nanomsg Next Gen) technology.
+#'     mirai is Japanese for 'future'.
 #'
 #' @section Links:
 #'
-#'     mirai website: \url{https://shikokuchuo.net/mirai/}
+#'     mirai website: \url{https://shikokuchuo.net/mirai/}\cr
+#'     mirai on CRAN: \url{https://cran.r-project.org/package=mirai}
 #'
 #'     nanonext website: \url{https://shikokuchuo.net/nanonext/}\cr
 #'     nanonext on CRAN: \url{https://cran.r-project.org/package=nanonext}
@@ -18,12 +18,30 @@
 #' @author Charlie Gao \email{charlie.gao@@shikokuchuo.net}
 #'     (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
 #'
-#' @importFrom nanonext call_aio context recv_ctx request send_aio send_ctx
-#'     socket stop_aio
+#' @importFrom nanonext call_aio context is_nul_byte recv_ctx request send_aio
+#'     send_ctx socket stop_aio unresolved .mirai_scm
 #' @importFrom stats runif
+#' @importFrom utils .DollarNames
 #'
 #' @docType package
 #' @name mirai-package
 #'
 NULL
+
+.onLoad <- function(libname, pkgname) {
+  daemons <- daemons()
+  daemons <<- daemons
+  invisible()
+}
+
+.onUnload <- function(libpath) {
+  daemons(0L)
+  invisible()
+}
+
+#' @export
+nanonext::is_nul_byte
+
+#' @export
+nanonext::unresolved