+++ tests/tests.R
@@ -1,11 +1,10 @@
 library(mirai)
 
-mirai <- eval_mirai(m + n + 1L, m = 2L, n = 3L)
-invisible(!is.null(mirai$aio) || stop())
-invisible(!is.null(mirai$socket) || stop())
-call_mirai(mirai)
-invisible(mirai$value == 6L || stop())
-invisible(is.null(mirai$aio) || stop())
-invisible(is.null(mirai$socket) || stop())
-invisible(identical(call_mirai(mirai), mirai) || stop())
+m <- eval_mirai({
+  q <- m + n + 1L
+  q / m
+}, m = 2L, n = 3L)
+invisible(inherits(call_mirai(m), "mirai") || stop())
+invisible(m$value == 3L || stop())
+invisible(identical(call_mirai(m), m) || stop())